+++ packages/kit/src/runtime/server/endpoint.js
@@ -1,4 +1,4 @@
-import { isContentTypeBinary } from '../../core/adapter-utils.js';
+import { isContentTypeTextual } from '../../core/adapter-utils.js';
 import { lowercase_keys } from './utils.js';
 
 /** @param {string} body */
@@ -8,6 +8,11 @@ function error(body) {
 		body,
 		headers: {}
 	};
+}
+
+/** @param {unknown} s */
+function is_string(s) {
+	return typeof s === 'string' || s instanceof String;
 }
 
 /**
@@ -38,18 +43,11 @@ export default async function render_route(request, route) {
 			headers = lowercase_keys(headers);
 			const type = headers['content-type'];
 
-			const is_type_binary = type && isContentTypeBinary(type);
+			const is_type_textual = isContentTypeTextual(type);
 
-			// validation
-			if (is_type_binary && !(body instanceof Uint8Array)) {
+			if (!is_type_textual && !(body instanceof Uint8Array || is_string(body))) {
 				return error(
-					`${preface}: body must be an instance of Uint8Array if content type is image/*, audio/*, video/* or application/octet-stream`
-				);
-			}
-
-			if (body instanceof Uint8Array && !is_type_binary) {
-				return error(
-					`${preface}: Uint8Array body must have content-type header of image/*, audio/*, video/* or application/octet-stream`
+					`${preface}: body must be an instance of string or Uint8Array if content-type is not a supported textual content-type`
 				);
 			}
 
@@ -63,7 +61,7 @@ export default async function render_route(request, route) {
 				(!type || type.startsWith('application/json'))
 			) {
 				headers = { ...headers, 'content-type': 'application/json; charset=utf-8' };
-				normalized_body = JSON.stringify(body || {});
+				normalized_body = JSON.stringify(typeof body === 'undefined' ? {} : body);
 			} else {
 				normalized_body = /** @type {import('types/hooks').StrictBody} */ (body);
 			}