+++ packages/kit/test/apps/basics/src/routes/endpoint-output/null.js
@@ -1,0 +1,4 @@
+/** @type {import('@sveltejs/kit').RequestHandler} */
+export function get() {
+	return { body: null };
+}