+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sveltejs/kit
 
+## 1.0.0-next.127
+
+### Patch Changes
+
+- bb3ae21: Fix endpoint validation to allow returning string for all content types
+
+## 1.0.0-next.126
+
+### Patch Changes
+
+- 4720b67: Default body parsing to binary
+- 6da07b8: fix returning null from endpoints
+
 ## 1.0.0-next.125
 
 ### Patch Changes